feat: find colors using regex instead of using `self.view.word()`
This prepares for alpha support. Alpha colors look like `rgba(1, 2, 3)`, which can't be extracted with `self.view.word()`. This also prepares support for `rgb(1, 2, 3)`, which similarly can't be extracted this way.

A nice side benefit, is that if I select the entire line, then all the colors in the line get replaced.